# workspace/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mini football game booking and player management platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /games/{gameId}/teams:
    get:
      operationId: getTeams
      tags: [teams]
      summary: Get current team split
      parameters:
        - $ref: "#/components/parameters/GameId"
      responses:
        "200":
          description: Teams
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamSplit"
    post:
      operationId: splitTeams
      tags: [teams]
      summary: Auto-split approved players into teams (admin)
      parameters:
        - $ref: "#/components/parameters/GameId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SplitTeamsBody"
      responses:
        "200":
          description: Teams created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamSplit"
  /games/{gameId}/teams/assignments:
    patch:
      operationId: moveTeamPlayer
      tags: [teams]
      summary: Move a player between teams (admin)
      parameters:
        - $ref: "#/components/parameters/GameId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveTeamPlayerBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamSplit"
...
components:
  parameters:
    GameId:
      name: gameId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    PublicPlayer:
      type: object
      required: [id, name, attendance]
      properties:
        id:
          type: string
        name:
          type: string
        attendance:
          $ref: "#/components/schemas/Attendance"
...
    Attendance:
      type: string
      enum: [confirmed, tentative]
...
    SplitTeamsBody:
      type: object
      required: [teamCount]
      properties:
        teamCount:
          type: integer
          enum: [2, 3]
    MoveTeamPlayerBody:
      type: object
      required: [registrationId, teamId]
      properties:
        registrationId:
          type: string
        teamId:
          type: string
    Team:
      type: object
      required: [id, name, color, players]
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
        players:
          type: array
          items:
            $ref: "#/components/schemas/PublicPlayer"
    TeamSplit:
      type: object
      required: [gameId, teams]
      properties:
        gameId:
          type: string
        teams:
          type: array
          items:
            $ref: "#/components/schemas/Team"
...